BSAS-9 [+] modify 'int id' -> 'Long id'

+++ engine/src/main/java/com/jsoftware/platform/repository/RecommendActivityRepository.java
@@ -10,7 +10,7 @@ import java.util.List;
 public interface RecommendActivityRepository {
     List<RecommendActivity> readRecommendActivities();
 
-    RecommendActivity readRecommendActivity(int id);
+    RecommendActivity readRecommendActivity(Long id);
 
     void createRecommendActivity(RecommendActivity recommendActivity);
 }
